feat: enhance role and permission management with updated API responses and seeder logic

+++ routes/api.php
@@ -56,42 +56,38 @@ Route::prefix('v1')->group(function () {
         });
 
 
-        Route::middleware('role:superadmin,admin')->group(function () {
+        Route::middleware('role:superadmin,admin,staff')->group(function () {
             // activity-logs
-            Route::prefix('activity-logs')->group(function () {
+            Route::prefix('activity-logs')->middleware('modules:Activity Logs')->group(function () {
                 Route::get('/', [ActivityLogController::class, 'showLogs'])->middleware('permissions:View All');
-                Route::get('/export', [ActivityLogController::class, 'export']);
+                Route::get('/export', [ActivityLogController::class, 'export'])->middleware('permissions:Export');
             });
             // Users
-            Route::prefix('users')->group(function () {
-                Route::get('/', [UserController::class, 'index']);
+            Route::prefix('users')->middleware('modules:Users')->group(function () {
+                Route::get('/', [UserController::class, 'index'])->middleware('permissions:View');
                 //modules middleware of users
-                Route::middleware('modules:Users')->group(function () {
-                    Route::get('/{id}', [UserController::class, 'show'])->middleware('permissions:View');
-                    Route::post('/', [UserController::class, 'store'])->middleware('permissions:Create');
-                    Route::post('/{id}/reset-password', [UserController::class, 'resetPassword'])->middleware('permissions:Edit');;
-                    Route::patch('/{id}', [UserController::class, 'update'])->middleware('permissions:Edit');
-                    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('permissions:Delete');
-                });
+                Route::get('/{id}', [UserController::class, 'show'])->middleware('permissions:View');
+                Route::post('/', [UserController::class, 'store'])->middleware('permissions:Create');
+                Route::post('/{id}/reset-password', [UserController::class, 'resetPassword'])->middleware('permissions:Edit');;
+                Route::patch('/{id}', [UserController::class, 'update'])->middleware('permissions:Edit');
+                Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('permissions:Delete');
             });
 
             // Roles
-            Route::prefix('roles')->group(function () {
-                Route::get('/', [RoleController::class, 'index']);
+            Route::prefix('roles')->middleware('modules:Roles')->group(function () {
+                Route::get('/', [RoleController::class, 'index'])->middleware('permissions:View');
 
                 //module middleware of roles
-                Route::middleware('modules:Roles')->group(function () {
-                    Route::get('/{id}', [RoleController::class, 'show'])->middleware('permissions:View');
-                    Route::post('/', [RoleController::class, 'store'])->middleware('permissions:Create');
-                    Route::put('/{id}', [RoleController::class, 'update'])->middleware('permissions:Edit');
-                    Route::delete('/{id}', [RoleController::class, 'destroy'])->middleware('permissions:Delete');
-                });
-                Route::post('/{role}/permissions', [RolePermissionController::class, 'assignPermissions']);
+                Route::get('/{id}', [RoleController::class, 'show'])->middleware('permissions:View');
+                Route::post('/', [RoleController::class, 'store'])->middleware('permissions:Create');
+                Route::put('/{id}', [RoleController::class, 'update'])->middleware('permissions:Edit');
+                Route::delete('/{id}', [RoleController::class, 'destroy'])->middleware('permissions:Delete');
+                Route::post('/{id}/permissions', [RolePermissionController::class, 'assignPermissions'])->middleware('permissions:Edit');
             });
 
             // Permissions
-            Route::prefix('permissions')->group(function () {
-                Route::get('/', [PermissionController::class, 'index']);
+            Route::prefix('permissions')->middleware('modules:Roles')->group(function () {
+                Route::get('/', [PermissionController::class, 'index'])->middleware('permissions:View');
             });
 
             // Categories
@@ -129,9 +125,9 @@ Route::prefix('v1')->group(function () {
                     Route::delete('/{id}', [AttributeController::class, 'destroy'])->middleware('permissions:Delete');
                 });
 
-                Route::post('/{id}/values', [AttributeController::class, 'storeAttributesValue']);
-                Route::patch('/values/{valueId}', [AttributeController::class, 'updateAttributesValue']);
-                Route::delete('/values/{valueId}', [AttributeController::class, 'destroyAttributesValue']);
+                Route::post('/{id}/values', [AttributeController::class, 'storeAttributesValue'])->middleware(['modules:Attributes', 'permissions:Create']);
+                Route::patch('/values/{valueId}', [AttributeController::class, 'updateAttributesValue'])->middleware(['modules:Attributes', 'permissions:Edit']);
+                Route::delete('/values/{valueId}', [AttributeController::class, 'destroyAttributesValue'])->middleware(['modules:Attributes', 'permissions:Delete']);
             });
 
             // Products
@@ -142,9 +138,9 @@ Route::prefix('v1')->group(function () {
                 Route::put('/{id}', [ProductController::class, 'update'])->middleware('permissions:Edit');
                 Route::delete('/{id}', [ProductController::class, 'destroy'])->middleware('permissions:Delete');
 
-                Route::get('/export', [ProductController::class, 'export']);
-                Route::post('/{id}/attributes/{attributeId}', [ProductController::class, 'storeAttribute']);
-                Route::delete('/{id}/attributeValueId/{attributeProductId}', [ProductController::class, 'destroyAttributeProduct']);
+                Route::get('/export', [ProductController::class, 'export'])->middleware('permissions:View');
+                Route::post('/{id}/attributes/{attributeId}', [ProductController::class, 'storeAttribute'])->middleware('permissions:Edit');
+                Route::delete('/{id}/attributeValueId/{attributeProductId}', [ProductController::class, 'destroyAttributeProduct'])->middleware('permissions:Edit');
             });
 
             //inventory
@@ -171,14 +167,14 @@ Route::prefix('v1')->group(function () {
             });
 
             // Stock-movements
-            Route::prefix('stock-movements')->group(function () {
+            Route::prefix('stock-movements')->middleware('role:superadmin,admin')->group(function () {
                 Route::get('/', [StocksController::class, 'showStockMovements']);
                 Route::get('/{id}', [StocksController::class, 'showStockMovementsById']);
                 Route::get('/cv/export', [StocksController::class, 'exportStockMovements']);
             });
 
             // Stock-adjustments
-            Route::prefix('stock-adjustments')->group(function () {
+            Route::prefix('stock-adjustments')->middleware('role:superadmin,admin')->group(function () {
                 Route::get('/', [StocksController::class, 'showStockAdjustments']);
                 Route::post('/', [StocksController::class, 'store']);
                 Route::get('/{id}', [StocksController::class, 'showStockAdjustmentsById']);
@@ -213,7 +209,7 @@ Route::prefix('v1')->group(function () {
 
 
             //Sales
-            Route::prefix('sales')->group(function () {
+            Route::prefix('sales')->middleware('role:superadmin,admin')->group(function () {
                 Route::get('/', [SalesController::class, 'index']);
                 Route::get('/{id}', [SalesController::class, 'show']);
                 Route::post('/{id}/void', [SalesController::class, 'void']);
@@ -222,13 +218,13 @@ Route::prefix('v1')->group(function () {
             });
 
             //Dashboard
-            Route::prefix('dashboard')->group(function () {
+            Route::prefix('dashboard')->middleware('modules:Dashboard')->group(function () {
                 Route::get('/charts/inventory-overview', [DashboardController::class, 'showInventoryOverview'])->middleware('permissions:View');
                 Route::get('/charts/revenue-by-category', [DashboardController::class, 'showRevenueByCategory'])->middleware('permissions:View');;
             });
 
             //Reports
-            Route::prefix('reports')->group(function () {
+            Route::prefix('reports')->middleware('modules:Reports')->group(function () {
                 Route::get('/sales', [ReportsController::class, 'showSalesReport'])->middleware('permissions:View');
                 Route::get('/purchases', [ReportsController::class, 'showPurchases'])->middleware('permissions:View');
                 Route::get('/inventory', [ReportsController::class, 'showInventory'])->middleware('permissions:View');
@@ -258,8 +254,8 @@ Route::prefix('v1')->group(function () {
 
 
         //Dashboard
-        Route::prefix('dashboard')->group(function () {
-            Route::get('/stats', [DashboardController::class, 'showStats']);
+        Route::prefix('dashboard')->middleware('modules:Dashboard')->group(function () {
+            Route::get('/stats', [DashboardController::class, 'showStats'])->middleware('permissions:View');
             Route::get('/charts/sales-trend', [DashboardController::class, 'showSalesTrend'])->middleware('permissions:View');
             Route::get('/charts/top-products', [DashboardController::class, 'showTopProducts'])->middleware('permissions:View');
             Route::get('/recent-activities', [DashboardController::class, 'showRecentActivities'])->middleware('permissions:View');
